Done
Was able to get the first to move data from one form to the next. Also
was able to hide the first form when I opened the next one. This can be
used to prevent access to the first form or it is possible to use
ShowDialog instead of this.Hide()